chore: rename /presentation/debits to /presentation/debts
